refactor with ElectricalRecording
