+++ tests/integration/test_searching.py
@@ -4,7 +4,9 @@ from entitysdk.models import (
     Contribution,
     ElectricalCellRecording,
     EModel,
+    IonChannel,
     IonChannelModel,
+    IonChannelRecording,
     License,
     MEModel,
     MEModelCalibrationResult,
@@ -26,7 +28,9 @@ from entitysdk.models import (
     "entity_type",
     [
         Contribution,
+        IonChannel,
         IonChannelModel,
+        IonChannelRecording,
         License,
         MTypeClass,
         Person,
